add config BLK_SNAP_FREE_DIFF_BUFFER_POOL_SIZE
A buffer can be allocated for each chunk. After use, this buffer is
not released immediately, but is sent to the pool of free buffers.
However, if there are too many free buffers in the pool, they are
released immediately. The maximum size of the pool is regulated by
this define.