add snapshot_mark_dirty_blocks()
To mark a block dirty, need to have the dev_id of the original device.
But this is not convenient for the user-space algorithm. Now we can
pass the dev_id of the snapshot image.
The module will determine that the major number corresponds to the
snapshot image and will search for the desired cbt table by the dev_id
of the image.