add option BDEV_FILTER_SYNC
The new build option BDEV_FILTER_SYNC allows you to make synchronous
bio requests in the filter.
Of course, the asynchronous mechanism looks more preferable, but it
does not guarantee that read and write requests will not be executed
in reverse order. This makes the idea of an asynchronous COW algorithm
impossible.