remove BDEV_FILTER_SYNC
Figured out how to work with flags REQ_SYNC, REQ_FUA and
REQ_PREFLUSH correctly. When reading and writing chunks, the use of
these flags guarantees the order in which bios are processed.
This made it possible to use current->bio_list when filtering bios,
which means to reduce the use of the stack to remove the blocking
of the write process for the duration of the COW algorithm.