add enum bdev_filter_result
Prior to this commit, the filter could only skip or pass the bio.
Using current->bio_list allows you to guarantee that there is no
recursion when calling submit_bio(). But using current->bio_list
does not allow to create a request, send it and wait for its
processing to be completed in one stage.
Now we can repeat processing the bio. This allows to perform bio
processing in several stages.