implementations of a filter with altitudes.
In this version of the filter, an array is used instead of a linked list
of filters. Each cell of the array corresponds to an altitude. For a
module, its own altitude is reserved in the tree. If the modules are not
compatible and cannot work together, then they share the same altitude.
Altitudes allow to guarantee the same sequence of bio processing for all
filters in the tree.
The "unidentified" altitude is reserved for out-of-tree filters.
I believe that the creators of out-of-tree modules will want to reserve
altitudes for their filter-modules. This motivate them to add their
modules to the kernels tree.

The read-write semaphore has also been removed. Now the array of filters
is protected from simultaneous access by a spinlock. And the filters
themselves have a reference counter, which allows to keep the filter
from being released as long as it is used.