altitude check has been added
Using enum does not completely protect against passing an incorrect
altitude parameter to the function. The altitude is an index for an
array, so in the case of an incorrect altitude, it will be quite
painful to catch an error going beyond the boundaries of the array.

But because of this check, the bdev_filter_get_by_altitude() function
began to return an error code. Handling of these errors has been added.