move bio_list_on_stack from bdevfilter to tracker
The problem is that if the COW algorithm is initiated, then it is
absolutely necessary to wait for the completion of reading data from
the original block device before overwriting. If we send two bios
in a device to read a sector, and then to write it, on some systems,
writing may outpace reading. Magic with flags REQ_SYNC and REQ_FUA does
not protect against this problem.
As a result, the interception algorithm changes. It becomes more
synchronous.
The good news is that this simplifies the code of the handling
algorithm.