circleci: add jobs for debian 10 and all supported ubuntu
This add jobs for all supported ubuntu LTS from 14.04 to 22.04 and 22.10
now in development, add also debian 10.
All docker images are the debian and ubuntu official ones.
For now only these are enabled by default: debian 10,11,12 and ubuntu
20.04,22,04,22,10

Note: ubuntu 14.04 need to enable backport repository and use it for
install cmake3 but I was not able to found the right way to add a
conditional run only for it with a fast try