fix device recreate problem
The peculiarity of the stand-alone version of the module is that it is
not possible to reliably intercept the block device shutdown event.
This led to the fact that when the block device was disconnected and
then connected, tracking was not reset.
Now a pointer to its structure is stored for a block device. If the
pointer has changed, then the device has been recreated, which means
tracking should be reset.