change logic for low_free_space event
The fix changes the time when the event about the request for a new
portion is sent.
Before the fix, the second portion was sent immediately after taking
the snapshot. Because of this, the second portion was always created,
even there was no activity on block devices under the snapshot.
After the correction, the second portion will be requested only when
the first portion is half full.