remove blksnap-dev
It turns out that in order to assemble the tool, we should first
build and install a package with libraries. Perhaps this is right,
but not particularly convenient. In addition, the tool is written in
such a way that it does not require compiled libraries. Only the header
file is required.