cross-platform handling logic
Previously, submit_bio_noacct_notrace was a variable with the address of
the submit_bio_noacсt() function, offset by 5 bytes. 5 bytes is the size
of the ftrace call command on x86 arch. Therefore, the code was
platform-dependent.
Now submit_bio_noacct_notrace() is a function. It overrides the value of
current->bio_list, it is equal to zero. Thus, it turns out to reveal
that the handling was performed from the submit_bio_noacсt_notrace().