allows to add a block device to only one snapshot
The module allows to create more than one snapshot at a time. Obviously,
adding one device to two different snapshots is an erroneous behavior.
The double addition of a block device to multiple snapshots caused an
error when releasing such snapshots. Now, when adding a block device to
a snapshot, a check is performed to ensure that the device has not yet
been added to the snapshot.