Merge with squash from stable-v2.0
I assume that the project will develop in the direction of supporting
the the in-tree Linux kernel blksnap module.

The standalone blksnap module distributed as part of the Veeam Agent
for Linux product will be located in branches with the prefix 'VAL'.

The history of patches for the kernel was saved in the 'patch'
directory.