add a first fio test "sequental_read"
The snapshot image is a virtual block device. I/O units to the image are
redirected to the original device. Therefore, the reading performance
from the snapshot image should be slightly lower than from the original
block device. It is important to know what the difference is. The test
allows to measure this difference.