leave the first 4KiB free in the difference storage
Quite often, the file system superblock is written to the difference
storage first. As a result, when scanning a block device with a
difference storage, a file system may be mistakenly detected on it.

The module does not change the contents of the first page, which will
allow to store metadata in it.